+++ src/viewers/PictureInPictureViewer.jsx
@@ -43,7 +43,7 @@ import { GLOBAL_SLIDER_DIMENSION_FIELDS } from '../constants';
  *     https://deck.gl/docs/developer-guide/interactivity#the-picking-info-object)
  * @param {Array} [props.transitionFields] A string array indicating which fields require a transition when making a new selection: Default: ['t', 'z'].
  * @param {function} [props.onViewportLoad] Function that gets called when the data in the viewport loads.
- * @param {Object} [props.glOptions] Additional options used when creating the WebGLContext.
+ * @param {Object} [props.deckProps] Additional options used when creating the DeckGL component.
  */
 
 const PictureInPictureViewer = props => {
@@ -71,7 +71,7 @@ const PictureInPictureViewer = props => {
     onHover,
     transitionFields = GLOBAL_SLIDER_DIMENSION_FIELDS,
     onViewportLoad,
-    glOptions
+    deckProps
   } = props;
   const {
     newselections,
@@ -139,7 +139,7 @@ const PictureInPictureViewer = props => {
       hoverHooks={hoverHooks}
       onViewStateChange={onViewStateChange}
       onHover={onHover}
-      glOptions={glOptions}
+      deckProps={deckProps}
     />
   );
 };